+++ packages/blocks/CHANGELOG.md
@@ -1,6 +1,24 @@
 # @cmpsr/blocks
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [465da90]
+  - @cmpsr/components@2.2.0
+
+## 1.1.0
+
+### Minor Changes
+
+- 1c35a47: feat: implement CookieBanner block
+
+### Patch Changes
+
+- 12049a9: feat: make the library public
+
 ## 1.0.0
+
 ### Patch Changes
 
 - Updated dependencies [0443122]